refactor(wire): one json module at the boundary
`http.ts` and `wire-boundary.ts` each held one vocabulary of the same
boundary `json.ts` already is, so a reader of the wire package had three
files to consult for what arrives from outside the build. The barrel's
exported names are unchanged.
